test(health): split part_001 under first-party line limit
test_health_part_001.py had grown to 1529 lines, over the 1500-line
first-party maintainability cap enforced by
test_first_party_code_files_stay_under_line_limit. Move the two trailing
self-contained tests (_cancel_unneeded_task, docker-check timeout) into
part_002, bringing part_001 to 1486 lines. Both parts are collected
during default discovery, so coverage is preserved.
